// src/api/generated/models/loginRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Arco Enterprise Admin API
 * Spring Boot 管理后台示例契约，建议由 springdoc-openapi 的 /v3/api-docs 替换。
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  remember?: boolean;
}